+++ explorer/src/lib/chain-info/chain-info.d.ts
@@ -37,14 +37,15 @@ type SearchResult = {
 type Transaction = {
   blockhash: string;
   blockheight: number;
-  contract: string;
   date: Date;
   feepaid: number;
   gaslimit: number;
   gasprice: number;
   gasspent: number;
+  memo: string;
   method: string;
   success: boolean;
   txerror: string;
   txid: string;
+  txtype: string;
 };