+++ explorer/src/lib/services/gql-queries.js
@@ -14,7 +14,10 @@ fragment TransactionInfo on SpentTransaction {
     },
     gasLimit,
     gasPrice,
-    id
+    id,
+    isDeploy,
+    memo
+    txType
   }
 }
 `;