+++ web-wallet/src/lib/mock-data/index.js
@@ -1,4 +1,7 @@
 export { default as addresses } from "./addresses";
 export { default as apiMarketData } from "./api-market-data.json";
+export { default as cacheHistory } from "./cache-history";
+export { default as cacheSpentNotes } from "./cache-spent-notes";
+export { default as cacheUnspentNotes } from "./cache-unspent-notes";
 export { default as stakeInfo } from "./stakeInfo";
 export { default as transactions } from "./transactions";