+++ web-wallet/src/lib/wallet-cache/wallet-cache.d.ts
@@ -1,0 +1,13 @@
+type WalletCacheNote = {
+  block_height: number;
+  note: number[];
+  nullifier: number[];
+  pos: number;
+  psk: string;
+};
+
+type WalletCacheHistoryEntry = {
+  history: Transaction[];
+  lastBlockHeight: number;
+  psk: string;
+};