+++ explorer/CHANGELOG.md
@@ -27,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add accessible name to the nav bar button on mobile [#2036]
 - Implement warning for stale market data [#1892]
 - Add tooltip to current and pending stake showing the exact amount [#2363]
+- Add `memo` and `isDeploy` fields in transactions [#2362]
+- Add `txType` fields in transactions and display it [#2347]
 
 ### Changed
 
@@ -39,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix Average Fee Paid label [#2057]
 - Fix list items alignment on mobile [#2056]
+- Fix shield icons used for tx type [#2389]
 
 ## [0.1.0] - 2024-07-24
 
@@ -62,7 +65,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2159]: https://github.com/dusk-network/rusk/issues/2159
 [#2220]: https://github.com/dusk-network/rusk/issues/2220
 [#2348]: https://github.com/dusk-network/rusk/issues/2348
+[#2362]: https://github.com/dusk-network/rusk/issues/2362
 [#2363]: https://github.com/dusk-network/rusk/issues/2363
+[#2363]: https://github.com/dusk-network/rusk/issues/2347
+[#2389]: https://github.com/dusk-network/rusk/issues/2389
 
 <!-- VERSIONS -->